pkg/trace/api: drop dependency on pkg/proto/pbgo by moving config endpoint out
Remaining:
     github.com/DataDog/datadog-agent/pkg/config/remote
     github.com/DataDog/datadog-agent/pkg/config/remote/data
     github.com/DataDog/datadog-agent/pkg/tagger
     github.com/DataDog/datadog-agent/pkg/tagger/collectors
     github.com/DataDog/datadog-agent/pkg/util/grpc
     github.com/DataDog/datadog-agent/pkg/util/http
     github.com/DataDog/datadog-agent/pkg/util/profiling
