+++ pkg/trace/api/api_test.go
@@ -7,7 +7,6 @@ package api
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -17,23 +16,18 @@ import (
 	"os"
 	"reflect"
 	"strconv"
-	"strings"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/proto/pbgo"
 	"github.com/DataDog/datadog-agent/pkg/trace/config"
-	"github.com/DataDog/datadog-agent/pkg/trace/config/features"
 	"github.com/DataDog/datadog-agent/pkg/trace/info"
 	"github.com/DataDog/datadog-agent/pkg/trace/pb"
 	"github.com/DataDog/datadog-agent/pkg/trace/sampler"
 	"github.com/DataDog/datadog-agent/pkg/trace/test/testutil"
-	"google.golang.org/grpc"
 
 	"github.com/cihub/seelog"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
 	"github.com/tinylib/msgp/msgp"
 	vmsgp "github.com/vmihailenco/msgpack/v4"
 )
@@ -60,42 +54,6 @@ func newTestReceiverFromConfig(conf *config.AgentConfig) *HTTPReceiver {
 	receiver := NewHTTPReceiver(conf, dynConf, rawTraceChan, noopStatsProcessor{})
 
 	return receiver
-}
-
-func newTestReceiverFromConfigWithGRPC(conf *config.AgentConfig, grpcClient pbgo.AgentSecureClient) *HTTPReceiver {
-	receiver := newTestReceiverFromConfig(conf)
-	receiver.coreClient = grpcClient
-	return receiver
-}
-
-type mockAgentSecureServer struct {
-	pbgo.AgentSecureClient
-	mock.Mock
-}
-
-func (a *mockAgentSecureServer) TaggerStreamEntities(ctx context.Context, in *pbgo.StreamTagsRequest, opts ...grpc.CallOption) (pbgo.AgentSecure_TaggerStreamEntitiesClient, error) {
-	args := a.Called(ctx, in, opts)
-	return args.Get(0).(pbgo.AgentSecure_TaggerStreamEntitiesClient), args.Error(1)
-}
-
-func (a *mockAgentSecureServer) TaggerFetchEntity(ctx context.Context, in *pbgo.FetchEntityRequest, opts ...grpc.CallOption) (*pbgo.FetchEntityResponse, error) {
-	args := a.Called(ctx, in, opts)
-	return args.Get(0).(*pbgo.FetchEntityResponse), args.Error(1)
-}
-
-func (a *mockAgentSecureServer) DogstatsdCaptureTrigger(ctx context.Context, in *pbgo.CaptureTriggerRequest, opts ...grpc.CallOption) (*pbgo.CaptureTriggerResponse, error) {
-	args := a.Called(ctx, in, opts)
-	return args.Get(0).(*pbgo.CaptureTriggerResponse), args.Error(1)
-}
-
-func (a *mockAgentSecureServer) DogstatsdSetTaggerState(ctx context.Context, in *pbgo.TaggerState, opts ...grpc.CallOption) (*pbgo.TaggerStateResponse, error) {
-	args := a.Called(ctx, in, opts)
-	return args.Get(0).(*pbgo.TaggerStateResponse), args.Error(1)
-}
-
-func (a *mockAgentSecureServer) ClientGetConfigs(ctx context.Context, in *pbgo.ClientGetConfigsRequest, opts ...grpc.CallOption) (*pbgo.ClientGetConfigsResponse, error) {
-	args := a.Called(ctx, in, opts)
-	return args.Get(0).(*pbgo.ClientGetConfigsResponse), args.Error(1)
 }
 
 func newTestReceiverConfig() *config.AgentConfig {
@@ -864,67 +822,6 @@ func TestClientComputedTopLevel(t *testing.T) {
 
 	t.Run("on", run(true))
 	t.Run("off", run(false))
-}
-
-func TestConfigEndpoint(t *testing.T) {
-	defer func(old string) { features.Set(old) }(strings.Join(features.All(), ","))
-
-	var tcs = []struct {
-		name               string
-		reqBody            string
-		expectedStatusCode int
-		enabled            bool
-		valid              bool
-		response           string
-	}{
-		{
-			name:               "disabled",
-			expectedStatusCode: http.StatusNotFound,
-			response:           "404 page not found\n",
-		},
-		{
-			name:               "bad",
-			enabled:            true,
-			expectedStatusCode: http.StatusBadRequest,
-			response:           "unexpected end of JSON input\n",
-		},
-		{
-			name:    "valid",
-			reqBody: `{"client":{"id":"test_client"}}`,
-
-			enabled:            true,
-			valid:              true,
-			expectedStatusCode: http.StatusOK,
-			response:           `{"targets":{"version":1,"raw":"dGVzdA=="}}`,
-		},
-	}
-	for _, tc := range tcs {
-		t.Run(tc.name, func(t *testing.T) {
-			assert := assert.New(t)
-			if tc.enabled {
-				features.Set("config_endpoint")
-			}
-			conf := newTestReceiverConfig()
-			grpc := mockAgentSecureServer{}
-			rcv := newTestReceiverFromConfigWithGRPC(conf, &grpc)
-			mux := rcv.buildMux()
-			server := httptest.NewServer(mux)
-			if tc.valid {
-				var request pbgo.ClientGetConfigsRequest
-				err := json.Unmarshal([]byte(tc.reqBody), &request)
-				assert.NoError(err)
-				grpc.On("ClientGetConfigs", mock.Anything, &request, mock.Anything).Return(&pbgo.ClientGetConfigsResponse{Targets: &pbgo.TopMeta{Version: 1, Raw: []byte("test")}}, nil)
-			}
-			req, _ := http.NewRequest("POST", server.URL+"/v0.7/config", strings.NewReader(tc.reqBody))
-			req.Header.Set("Content-Type", "application/msgpack")
-			resp, err := http.DefaultClient.Do(req)
-			assert.Nil(err)
-			body, err := ioutil.ReadAll(resp.Body)
-			assert.Nil(err)
-			assert.Equal(tc.expectedStatusCode, resp.StatusCode)
-			assert.Equal(tc.response, string(body))
-		})
-	}
 }
 
 func TestClientDropP0s(t *testing.T) {
