+++ pkg/trace/api/version.go
@@ -80,5 +80,5 @@ const (
 	// Payload: TracerPayload (pkg/trace/pb/tracer_payload.proto)
 	// Response: Service sampling rates.
 	//
-	v07 Version = "v0.7"
+	V07 Version = "v0.7"
 )
